requirements.txt
arrow==1.2.3
boto3==1.26.78
botocore==1.29.78
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
make==0.1.6.post2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
pydotenv==0.0.7
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.3
